update config yaml file
